requirements.txt
splat64[mips]==0.31.0
spimdisasm==1.31.3
ninja_syntax>=1.7.2
argparse
GitPython
requests
mapfile_parser
n64img
pypng
